chore(broker): extract exclusive gateway test
